+++ driver/hpm/hpm_canfd.cpp
@@ -1,4 +1,4 @@
-﻿#include "hpm_mcan.hpp"
+#include "hpm_canfd.hpp"
 
 #include <cstring>
